LC-path-sum-iii [Runtime: N/A; Memory: N/A]

+++ binary-tree/path-sum-iii.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    int prefixSum = 0;
+    long long prefixSum = 0;
     unordered_map<long long, long long> count;
     
     int ans = 0;
